--- src/dep_solver.cpp
@@ -1,400 +1,0 @@
-/**************************************************/
-/* Dependency solver for the birb package manager */
-/**************************************************/
-
-#include "Birb.hpp"
-#include <algorithm>
-#include <cassert>
-#include <filesystem>
-#include <fstream>
-#include <iostream>
-#include <string.h>
-#include <string>
-#include <unordered_map>
-#include <unordered_set>
-#include <vector>
-
-/* Function declarations */
-std::vector<std::string> get_dependencies(const std::string& pkg, const std::vector<pkg_source>& repos, int depth);
-std::vector<std::string> get_reverse_dependencies(const std::string& pkg_name, const std::vector<pkg_source>& repos);
-std::vector<std::string> deduplicated_dep_list(const std::vector<std::string>& dependencies);
-void orphan_finder(const std::vector<pkg_source>& repos);
-
-
-static std::unordered_map<std::string, std::vector<std::string>> meta_packages;
-static std::unordered_map<std::string, std::vector<std::string>> dependency_cache;
-static std::unordered_map<std::string, std::vector<std::string>> reverse_dependency_cache;
-
-enum Mode
-{
-	DEPENDENCY_SOLVER,
-	REVERSE_DEPENDENCY_SOLVER,
-	ORPHAN_FINDER,
-};
-
-
-std::vector<std::string> get_dependencies(const std::string& pkg, const std::vector<pkg_source>& repos, int depth)
-{
-	assert(pkg.empty() == false);
-	assert(repos.size() > 0);
-
-	/* Check if this package has its dependencies in the cache already */
-	if (dependency_cache.contains(pkg))
-		return dependency_cache[pkg];
-
-	std::vector<std::string> deps;
-
-	/* Avoid dependency loops and infinite (or unnecessary) recursion */
-	if (depth < 0)
-		return deps;
-
-	/* Find the repo that has the package */
-	pkg_source repo = birb::locate_pkg_repo(pkg, repos);
-
-	/* Return empty dependency list if the package couldn't be found */
-	if (!repo.is_valid())
-		return deps;
-
-	/* Perform string splitting only if this isn't a meta package */
-	if (!meta_packages.contains(pkg))
-	{
-		/* Read data from the package file */
-		std::string dep_line = birb::read_pkg_variable(pkg, "DEPS", repo.path);
-
-		/* Return empty dependency list if the dep_line is empty
-		 * The line could be empty for example when the package doesn't exist or is invalid etc. */
-		if (dep_line.empty())
-			return deps;
-
-		/* Split the string */
-		size_t pos = 0;
-		while ((pos = dep_line.find(" ")) != std::string::npos)
-		{
-			std::string dep = dep_line.substr(0, pos);
-
-			/* Check if the dependency is a meta package and should be expanded */
-			if (meta_packages.contains(dep))
-			{
-				/* Expand the meta package */
-				deps.insert(deps.end(), meta_packages[dep].begin(), meta_packages[dep].end());
-			}
-			else
-			{
-				deps.push_back(dep);
-			}
-
-			dep_line.erase(0, pos + 1);
-		}
-
-		if (!dep_line.empty())
-			deps.push_back(dep_line);
-
-		assert(deps.size() > 0);
-	}
-	else
-	{
-		deps = meta_packages[pkg];
-	}
-
-	/* Get dependencies of the dependencies of this package, if any */
-	size_t deps_size = deps.size();
-	for (size_t i = 0; i < deps_size; ++i)
-	{
-		std::vector<std::string> sub_deps = get_dependencies(deps[i], repos, depth - 1);
-		deps.insert(deps.end(), sub_deps.begin(), sub_deps.end());
-	}
-
-	assert(dependency_cache.contains(pkg) == false && "Overwriting old cache results");
-
-	/* Cache the results */
-	dependency_cache[pkg] = deps;
-
-	return deps;
-}
-
-std::vector<std::string> get_reverse_dependencies(const std::string& pkg_name, const std::vector<pkg_source>& repos)
-{
-	std::vector<std::string> dependencies;
-
-	/* Get list of installed packages */
-	std::vector<std::string> installed_packages = birb::get_installed_packages();
-
-	/* Get the reverse dependencies for the package with no recursion */
-	std::vector<std::string> temp_deps;
-	for (size_t i = 0; i < installed_packages.size(); ++i)
-	{
-		temp_deps = get_dependencies(installed_packages[i], repos, 0);
-
-		/* Check if the package had this package we are inspecting in
-		 * its dependency list */
-		if (std::find(temp_deps.begin(), temp_deps.end(), pkg_name) != temp_deps.end())
-			dependencies.push_back(installed_packages[i]);
-	}
-
-	return dependencies;
-}
-
-std::vector<std::string> deduplicated_dep_list(const std::vector<std::string>& dependencies)
-{
-	std::vector<std::string> result;
-
-	/* Don't do anything if there are no dependencies */
-	if (dependencies.empty())
-		return result;
-
-	std::unordered_set<std::string> exists_in_results;
-
-	/* Start from the end of the list and add each package
-	 * to the list only once */
-	for (int i = dependencies.size() - 1; i >= 0; --i)
-	{
-		/* Skip meta-packages */
-		if (meta_packages.contains(dependencies[i]))
-			continue;
-
-		/* Check if the package is already in the result list */
-		if (exists_in_results.contains(dependencies[i]))
-			continue;
-
-		result.push_back(dependencies[i]);
-		exists_in_results.insert(dependencies[i]);
-	}
-
-	return result;
-}
-
-void orphan_finder(const std::vector<pkg_source>& repos)
-{
-	std::unordered_set<std::string> result;
-
-	/* Read in the list of packages installed by the user */
-	std::vector<std::string> nest = birb::read_file("/var/lib/birb/nest");
-
-	/* Get the list of all installed applications */
-	std::vector<std::string> installed_packages = birb::get_installed_packages();
-
-	/* Orphan candidates are packages that are installed but aren't in the nest */
-	std::vector<std::string> orphan_candidates;
-	orphan_candidates.reserve(installed_packages.size() - nest.size());
-
-	/* Find all orphan candidates */
-	for (size_t i = 0; i < installed_packages.size(); ++i)
-	{
-		if (std::find(nest.begin(), nest.end(), installed_packages[i]) == nest.end())
-			orphan_candidates.push_back(installed_packages[i]);
-	}
-	std::cout << "Found " << orphan_candidates.size() << " orphan candidates\n";
-
-	/* Call it quits if there are no packages that could be orphans */
-	if (orphan_candidates.size() == 0)
-		return;
-
-	std::cout << "Starting the orphan scan\n";
-
-	pkg_source repo;
-	std::string flags_var;
-	std::vector<std::string> flags;
-	std::vector<std::string> reverse_deps;
-	std::vector<std::string> clean_reverse_deps;
-
-	constexpr int max_passes = 256;
-	bool clean_run = true;
-
-	for (int pass = 0; pass < max_passes; ++pass)
-	{
-		clean_run = true;
-		for (size_t i = 0; i < orphan_candidates.size(); ++i)
-		{
-			std::cout << "\rScanning... [Pass: " << pass + 1 << "] [" << i + 1 << "/" << orphan_candidates.size() << "]" << std::flush;
-
-			/* Skip packages that are already in the result list */
-			if (result.contains(orphan_candidates[i]))
-				continue;
-
-			/* Skip the package if it doesn't have a fakeroot */
-			if (!std::filesystem::exists("/var/db/fakeroot/" + orphan_candidates[i]))
-				continue;
-
-			/* Check if the package is "important" and should be skipped */
-			repo      = birb::locate_pkg_repo(orphan_candidates[i], repos);
-			flags_var = birb::read_pkg_variable(orphan_candidates[i], "FLAGS", repo.path);
-
-			if (!flags_var.empty())
-			{
-				flags = birb::split_string(flags_var, " ");
-				if (std::find(flags.begin(), flags.end(), "important") !=  flags.end())
-					continue;
-			}
-
-			/* Check if the package has any reverse dependencies */
-			if (reverse_dependency_cache.contains(orphan_candidates[i]))
-			{
-				reverse_deps = reverse_dependency_cache[orphan_candidates[i]];
-			}
-			else
-			{
-				reverse_deps = get_reverse_dependencies(orphan_candidates[i], repos);
-				reverse_dependency_cache[orphan_candidates[i]] = reverse_deps;
-			}
-			clean_reverse_deps.clear();
-
-			/* Clean out packages that have already been confirmed to be orphans */
-			for (size_t j = 0; j < reverse_deps.size(); ++j)
-			{
-				if (result.contains(reverse_deps[j]))
-					continue;
-
-				clean_reverse_deps.push_back(reverse_deps[j]);
-			}
-
-			if (clean_reverse_deps.empty())
-			{
-				result.insert(orphan_candidates[i]);
-				clean_run = false;
-			}
-		}
-
-		/* We found all orphans */
-		if (clean_run)
-			break;
-	}
-	std::cout << "\n";
-
-	/* Reverse the result set */
-	std::vector<std::string> orphans(result.size());
-	int i = orphans.size() - 1;
-	for (std::string p : result)
-	{
-		orphans[i] = p;
-		--i;
-	}
-
-	for (std::string o : orphans)
-		std::cerr << o << " ";
-
-	std::cout << "\n";
-}
-
-int main(int argc, char** argv)
-{
-	/* Name of the package that we are inspecting */
-	std::string pkg_name;
-
-	/* Location where we can find all of the packages in */
-	std::vector<pkg_source> repos;
-
-	/* Dependency resolution direction */
-	Mode operation_mode = DEPENDENCY_SOLVER;
-
-	/* Testing args */
-	if (argc == 4)
-	{
-		if (!strcmp(argv[1], "--test"))
-		{
-			repos.clear();
-
-			repos.push_back(pkg_source("Testing repo", "0.0.0.0", argv[2]));
-			pkg_name 		= argv[3];
-		}
-	}
-	else if (argc == 3) /* Some extra argument + package name */
-	{
-		if (!strcmp(argv[1], "-r") || !strcmp(argv[1], "--reverse"))
-		{
-			operation_mode = REVERSE_DEPENDENCY_SOLVER;
-		}
-		else /* No valid argument was given */
-		{
-			std::cout << "Invalid argument: " << argv[1] << "\n";
-			exit(1);
-		}
-
-		pkg_name = argv[2];
-		repos = birb::get_pkg_sources();
-	}
-	else if (argc == 2)
-	{
-		if (!strcmp(argv[1], "-o") || !strcmp(argv[1], "--orphan"))
-		{
-			operation_mode = ORPHAN_FINDER;
-		}
-		else
-		{
-			/* Only the package name was provided */
-			pkg_name = argv[1];
-		}
-
-		repos = birb::get_pkg_sources();
-	}
-	else
-	{
-		std::cout << "Incorrect amount of arguments!\n";
-		return 1;
-	}
-
-	/* Read in the list of meta-packages and parse them into a map */
-	std::vector<std::string> meta_package_list;
-
-	/* Loop through all repos to get all meta_packages */
-	assert(repos.size() > 0);
-	std::vector<std::string> tmp_meta;
-	std::string meta_path;
-	for (pkg_source s : repos)
-	{
-		assert(s.path.empty() == false);
-		assert(s.name.empty() == false);
-		assert(s.url.empty()  == false);
-		assert(s.is_valid()   == true);
-
-		/* Check if the repo has a meta_package file */
-		meta_path = s.path + "/meta_packages";
-		if (!std::filesystem::exists(meta_path))
-			continue;
-
-		tmp_meta = birb::read_file(meta_path);
-		meta_package_list.insert(meta_package_list.end(), tmp_meta.begin(), tmp_meta.end());
-	}
-
-	for (std::string line : meta_package_list)
-	{
-		/* Find the delimiter */
-		size_t pos = line.find(":");
-
-		/* Skip the line if the delimiter couldn't be found */
-		if (pos == std::string::npos)
-			continue;
-
-		/* Get the key and the corresponding value and assign it into a map */
-		meta_packages[line.substr(0, pos)] = birb::split_string(line.substr(pos + 1, line.size() - (pos + 1)), " ");
-	}
-
-	std::vector<std::string> dependencies;
-
-	switch (operation_mode)
-	{
-		/* Resolve dependencie normally for the given package */
-		case (DEPENDENCY_SOLVER):
-			/* Get all package dependencies recursively */
-			dependencies = get_dependencies(pkg_name, repos, 512);
-
-			/* Deduplicate the dependency list */
-			dependencies = deduplicated_dep_list(dependencies);
-			break;
-
-
-		/* Find packages that depend on this given package */
-		case (REVERSE_DEPENDENCY_SOLVER):
-			dependencies = get_reverse_dependencies(pkg_name, repos);
-			break;
-
-		case (ORPHAN_FINDER):
-			orphan_finder(repos);
-			return 0;
-			break;
-	};
-
-	for (std::string d : dependencies)
-		std::cout << d << "\n";
-
-	return 0;
-}
