+++ src/frontend/pkg_search.cpp
